+++ src/frontend/components/CatchAllUrlRule.php
@@ -55,7 +55,7 @@ class CatchAllUrlRule extends UrlRule
         }
         
         // return the custom route
-        return ['/cms/default/index', ['path' => $request->pathInfo]];
+        return [$this->route, ['path' => $request->pathInfo]];
     }
 
     /**
@@ -63,7 +63,7 @@ class CatchAllUrlRule extends UrlRule
      */
     public function createUrl($manager, $route, $params)
     {
-        if (ltrim($route, '/') !== 'cms/default/index') {
+        if (ltrim($route, '/') !== $this->route) {
             $this->createStatus = self::CREATE_STATUS_ROUTE_MISMATCH;
             return false;
         }